+++ bin/autodoc.sh
@@ -1,8 +1,0 @@
-#!/usr/bin/env bash
-
-lein autodoc
-cd autodoc
-git add .
-git commit -m "Documentation updated"
-git push
-cd ..